+++ Sample/Sample/HostViewController.swift
@@ -27,21 +27,22 @@ class HostViewController: MenuContainerViewController {
     
     override func viewDidLoad() {
         super.viewDidLoad()
-     
-        menuViewController = self.storyboard!.instantiateViewController(withIdentifier: "NavigationMenu") as! MenuViewController
-        
-        contentViewControllers = contentControllers()
+        let screenSize: CGRect = UIScreen.main.bounds
+        self.transitionOptions = TransitionOptions(duration: 0.4, visibleContentWidth: screenSize.width / 6)
 
+        menuViewController = self.storyboard!.instantiateViewController(withIdentifier: "NavigationMenu") as! MenuViewController
+        contentViewControllers = contentControllers()
         selectContentViewController(contentViewControllers.first!)
     }
-    
-    override func menuTransitionOptionsBuilder() -> TransitionOptionsBuilder? {
-        return TransitionOptionsBuilder() { builder in
-            builder.duration = 0.6
-            builder.contentScale = 0.9
-        }
+
+    override func viewWillTransition(to size: CGSize, with coordinator: UIViewControllerTransitionCoordinator) {
+        super.viewWillTransition(to: size, with: coordinator)
+        var options = TransitionOptions()
+        options.duration = size.width < size.height ? 0.4 : 0.6
+        options.visibleContentWidth = size.width / 6
+        self.transitionOptions = options
     }
-    
+
     private func contentControllers() -> [UIViewController] {
         let controllersIdentifiers = ["Kitty", "TabBar"]
         var contentList = [UIViewController]()